Remove `include/grpc/impl/codegen/compression_types.h`
