+++ include/grpc/compression.h
@@ -23,7 +23,7 @@
 
 #include <stdlib.h>
 
-#include <grpc/impl/codegen/compression_types.h>  // IWYU pragma: export
+#include <grpc/impl/compression_types.h>  // IWYU pragma: export
 #include <grpc/slice.h>
 
 #ifdef __cplusplus
