+++ include/grpcpp/client_context.h
@@ -38,8 +38,8 @@
 #include <memory>
 #include <string>
 
-#include <grpc/impl/codegen/compression_types.h>
 #include <grpc/impl/codegen/propagation_bits.h>
+#include <grpc/impl/compression_types.h>
 #include <grpcpp/impl/codegen/core_codegen_interface.h>
 #include <grpcpp/impl/codegen/create_auth_context.h>
 #include <grpcpp/impl/codegen/metadata_map.h>
