+++ include/grpcpp/impl/call_op_set.h
@@ -23,8 +23,8 @@
 #include <map>
 #include <memory>
 
-#include <grpc/impl/codegen/compression_types.h>
 #include <grpc/impl/codegen/grpc_types.h>
+#include <grpc/impl/compression_types.h>
 #include <grpc/support/log.h>
 #include <grpcpp/client_context.h>
 #include <grpcpp/completion_queue.h>
