+++ src/core/ext/transport/chttp2/transport/hpack_encoder.h
@@ -30,7 +30,7 @@
 #include "absl/strings/match.h"
 #include "absl/strings/string_view.h"
 
-#include <grpc/impl/codegen/compression_types.h>
+#include <grpc/impl/compression_types.h>
 #include <grpc/slice.h>
 #include <grpc/status.h>
