+++ src/core/lib/surface/call_test_only.h
@@ -23,8 +23,8 @@
 
 #include <stdint.h>
 
-#include <grpc/impl/codegen/compression_types.h>
 #include <grpc/impl/codegen/grpc_types.h>
+#include <grpc/impl/compression_types.h>
 
 /** Return the message compression algorithm from \a call.
  *
