+++ src/core/lib/surface/channel.h
@@ -36,8 +36,8 @@
 
 #include <grpc/event_engine/event_engine.h>
 #include <grpc/event_engine/memory_allocator.h>
-#include <grpc/impl/codegen/compression_types.h>
 #include <grpc/impl/codegen/grpc_types.h>
+#include <grpc/impl/compression_types.h>
 #include <grpc/slice.h>
 
 #include "src/core/lib/channel/channel_args.h"
