+++ src/cpp/common/channel_arguments.cc
@@ -20,8 +20,8 @@
 #include <string>
 #include <vector>
 
-#include <grpc/impl/codegen/compression_types.h>
 #include <grpc/impl/codegen/grpc_types.h>
+#include <grpc/impl/compression_types.h>
 #include <grpc/support/log.h>
 #include <grpcpp/grpcpp.h>
 #include <grpcpp/resource_quota.h>
