+++ src/cpp/server/server_builder.cc
@@ -28,8 +28,8 @@
 #include <vector>
 
 #include <grpc/grpc.h>
-#include <grpc/impl/codegen/compression_types.h>
 #include <grpc/impl/codegen/grpc_types.h>
+#include <grpc/impl/compression_types.h>
 #include <grpc/support/log.h>
 #include <grpc/support/sync.h>
 #include <grpc/support/workaround_list.h>
